+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.8.0] - 2024-11-20
+
+### Added
+
+- Add Customizable Working Directory setting to set the working directory for the extension
+- Add version extension check to show a message when the extension is outdated
+
 ## [2.7.0] - 2024-11-07
 
 ### Added
@@ -284,7 +291,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[unreleased]: https://github.com/ManuelGil/vscode-angular-generator/compare/v2.6.0...HEAD
+[unreleased]: https://github.com/ManuelGil/vscode-angular-generator/compare/v2.8.0...HEAD
+[2.8.0]: https://github.com/ManuelGil/vscode-angular-generator/compare/v2.7.0...v2.8.0
 [2.7.0]: https://github.com/ManuelGil/vscode-angular-generator/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/ManuelGil/vscode-angular-generator/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/ManuelGil/vscode-angular-generator/compare/v2.4.1...v2.5.0